fix: add dist to tracked files
In order to be able to try the package in another project, I decided to include the dist directory in tracked files. Ideally an npm package could be prepared and sent to some npm registry but to keep it simple I chose this drawback to just include dist in the repository.